split model and trainer
